LPS-160668 Add Confirm Deletion alert

+++ modules/apps/site-initializer/site-initializer-testray/extra/remote-app/src/services/rest/index.ts
@@ -12,10 +12,13 @@
  * details.
  */
 
+import i18n from '../../i18n';
 import fetcher from '../fetcher';
 
 const deleteResource = (resource: RequestInfo) => {
-	return fetcher.delete(resource);
+	if (confirm(i18n.translate('are-you-sure-you-want-to-delete-this-item'))) {
+		return fetcher.delete(resource);
+	}
 };
 
 export {deleteResource};
